requirements.txt
absl-py==0.12.0
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dataclasses==0.8
fsspec==2021.4.0
future==0.18.2
google-auth==1.29.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
idna==2.10
importlib-metadata==4.0.1
imutils==0.5.4
kiwisolver==1.3.1
Markdown==3.3.4
matplotlib==3.3.4
munch==2.5.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.5.1.48
Pillow==8.2.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytorch-lightning==1.0.8
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.7.0
torchfile==0.1.0
torchvision==0.8.1
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
